+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased/Snapshot]
 
+### Added
+- Introduced squants [#740](https://github.com/ie3-institute/PowerSystemUtils/issues/740)
+
 ## Changed
 - Cleaned up dependencies [#601](https://github.com/ie3-institute/PowerSystemUtils/issues/601)